7.4.修改各种bug,增加修改车辆功能，修改日期与年限选择，修改畅行无忧选择数量、限制时间----徐俊全

+++ ruyiruyiios/BuyTyre/View/OderBottomView.m
@@ -113,7 +113,7 @@
     self.tireTotalPriceLabel.text = [NSString stringWithFormat:@"¥ %@", tireTotalPriceStr];
 //    NSLog(@"%@", self.tireTotalPriceLabel.text);
     self.cxwyTotalPriceLabel.text = [NSString stringWithFormat:@"¥ %@", cxwyTotalPriceStr];
-    self.totalPriceLabel.text = [NSString stringWithFormat:@"¥ %.2f", ([tireTotalPriceStr floatValue] + [cxwyTotalPriceStr floatValue])];
+    self.totalPriceLabel.text = [NSString stringWithFormat:@"¥ %ld", (NSInteger)([tireTotalPriceStr floatValue] + [cxwyTotalPriceStr floatValue])];
 }
 
 /*
